% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_anosim.R
\name{print.nested_anosim}
\alias{print.nested_anosim}
\title{Print method for "nested_anosim" objects}
\usage{
\method{print}{nested_anosim}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class "nested_anosim".}

\item{digits}{Number of digits to print.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for "nested_anosim" objects
}
