% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_simper.R
\name{print.nested_simper}
\alias{print.nested_simper}
\title{Print method for "nested_simper" objects}
\usage{
\method{print}{nested_simper}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class "nested_simper".}

\item{digits}{Number of digits to print.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for "nested_simper" objects
}
