% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mistletoe_visitors}
\alias{mistletoe_visitors}
\title{Animal visitors to two mistletoe species at Las Chinchillas National Reserve}
\format{
A data frame with 87 rows and 43 columns (3 metadata columns and 40 visitor taxa):
\describe{
\item{Year_study}{Sampling year: \code{"First"} or \code{"Second"}.}
\item{Mistletoe}{Mistletoe species: \code{"T_aphyllus"} or \code{"T_verticillatus"}.}
\item{Camera}{Camera (mistletoe individual) code. The same cameras were
used in both years.}
\item{Anairetes_parulus}{Records of \emph{Anairetes parulus}.}
\item{butterfly}{Records of butterfly.}
\item{Callipepla_californica}{Records of \emph{Callipepla californica}.}
\item{Callopistes_maculatus}{Records of \emph{Callopistes maculatus}.}
\item{Cicadidae_sp}{Records of \emph{Cicadidae} sp.}
\item{Colorhamphus_parvirostris}{Records of \emph{Colorhamphus parvirostris}.}
\item{cricket}{Records of cricket.}
\item{Curaeus_curaeus}{Records of \emph{Curaeus curaeus}.}
\item{Diptera}{Records of Diptera.}
\item{Diuca_diuca}{Records of \emph{Diuca diuca}.}
\item{Elaenia_albiceps}{Records of \emph{Elaenia albiceps}.}
\item{Falco_sparverius}{Records of \emph{Falco sparverius}.}
\item{Geositta_cunicularia}{Records of \emph{Geositta cunicularia}.}
\item{Geranoaetus_polyosoma}{Records of \emph{Geranoaetus polyosoma}.}
\item{grasshopper}{Records of grasshopper.}
\item{Hemiptera}{Records of Hemiptera.}
\item{Hymenoptera}{Records of Hymenoptera.}
\item{Leistes_loyca}{Records of \emph{Leistes loyca}.}
\item{Leptasthenura_aegithaloides}{Records of \emph{Leptasthenura aegithaloides}.}
\item{Liolaemus_sp}{Records of \emph{Liolaemus} sp.}
\item{Lizard}{Records of Lizard.}
\item{Lycalopex_griseus}{Records of \emph{Lycalopex griseus}.}
\item{Molothrus_bonariensis}{Records of \emph{Molothrus bonariensis}.}
\item{moth}{Records of moth.}
\item{Muscisaxicola_maclovianus}{Records of \emph{Muscisaxicola maclovianus}.}
\item{Nothoprocta_perdicaria}{Records of \emph{Nothoprocta perdicaria}.}
\item{Phrygilus_alaudinus}{Records of \emph{Phrygilus alaudinus}.}
\item{Phrygilus_atriceps}{Records of \emph{Phrygilus atriceps}.}
\item{Phrygilus_gayi}{Records of \emph{Phrygilus gayi}.}
\item{Phyllotys_darwini}{Records of \emph{Phyllotys darwini}.}
\item{Porphyrospiza_alaudina}{Records of \emph{Porphyrospiza alaudina}.}
\item{Pseudasthenes_humicola}{Records of \emph{Pseudasthenes humicola}.}
\item{Pteroptochos_megapodius}{Records of \emph{Pteroptochos megapodius}.}
\item{Pyrisitia_lisa}{Records of \emph{Pyrisitia lisa}.}
\item{Pyrope_pyrope}{Records of \emph{Pyrope pyrope}.}
\item{Thylamys_elegans}{Records of \emph{Thylamys elegans}.}
\item{Troglodytes_aedon}{Records of \emph{Troglodytes aedon}.}
\item{Veniliornis_lignarius}{Records of \emph{Veniliornis lignarius}.}
\item{Zenaida_auriculata}{Records of \emph{Zenaida auriculata}.}
\item{Zonotrichia_capensis}{Records of \emph{Zonotrichia capensis}.}
}
}
\source{
Las Chinchillas National Reserve camera-trap project. See the
original study at \doi{10.1016/j.jaridenv.2025.105518}. A semicolon-separated
copy is available at
\code{system.file("extdata", "mistletoe_visitors.csv", package = "nestedANOSIM")}.
Please cite the paper if you use these data.
}
\usage{
mistletoe_visitors
}
\description{
Camera-trap records of animal visitors to the mistletoes \emph{Tristerix
aphyllus} and \emph{T. verticillatus} in Las Chinchillas National Reserve,
northern Chile, over two sampling years. Each row is one camera
(one mistletoe individual) in one sampling year. Counts are the number of
records of each taxon.
}
\details{
The focal species (\emph{Mimus thenca} and \emph{Sephanoides sephaniodes}) and
unidentified records were removed, as in the original analysis. Seven
cameras recorded none of the remaining taxa; they are kept so that the
automatic removal of empty samples in \code{\link[=nested_anosim]{nested_anosim()}} can be seen.
}
\examples{
data(mistletoe_visitors)
table(mistletoe_visitors$Year_study, mistletoe_visitors$Mistletoe)
comm <- mistletoe_visitors[, -(1:3)]
sort(colSums(comm), decreasing = TRUE)[1:5]
}
\keyword{datasets}
